+++ src/commands/ps.rs
@@ -1,6 +1,5 @@
 use anyhow::Result;
 use clap::ValueEnum;
-use indexmap::IndexSet;
 use itertools::Itertools;
 
 use crate::{
@@ -60,53 +59,63 @@ pub(crate) async fn run(args: Args, config: Config) -> Result<()> {
     let file = compose::parse(&config, false)?;
     let name = file.name.as_ref().unwrap();
 
-    if args.quiet {
-        print!(
-            "{}",
-            podman
-                .run(
-                    [
-                        "ps",
-                        "--quiet",
-                        "--no-trunc",
-                        "--filter",
-                        &format!("pod={name}")
-                    ]
-                    .into_iter()
-                    .chain(if args.all { vec!["--all"] } else { vec![] }),
-                )
-                .await?
-        );
-    } else if args.service {
-        let output = podman
-            .run(
-                ["ps", "--format", "json", "--filter", &format!("pod={name}")]
-                    .into_iter()
-                    .chain(if args.all { vec!["--all"] } else { vec![] }),
-            )
-            .await?;
-        let services = serde_json::from_str::<Vec<Container>>(&output)?
-            .into_iter()
-            .filter_map(|container| {
-                container
-                    .labels
-                    .and_then(|labels| labels.service)
-                    .filter(|service| {
-                        args.services.contains(service)
-                            || (args.services.is_empty() && file.services.keys().contains(service))
-                    })
-            })
-            .collect::<IndexSet<_>>();
+    let filters = args
+        .status
+        .into_iter()
+        .map(|status| format!("status={status:?}").to_ascii_lowercase())
+        .chain(args.filter)
+        .collect::<Vec<_>>();
+    let output = podman
+        .force_run(
+            ["ps", "--format", "json", "--filter", &format!("pod={name}")]
+                .into_iter()
+                .chain(if args.all { vec!["--all"] } else { vec![] })
+                .chain(filters.iter().flat_map(|filter| ["--filter", filter])),
+        )
+        .await?;
+    let containers = serde_json::from_str::<Vec<Container>>(&output)?
+        .into_iter()
+        .filter_map(|container| {
+            container
+                .labels
+                .as_ref()
+                .and_then(|labels| labels.service.clone())
+                .and_then(|service| {
+                    if args.services.contains(&service)
+                        || (args.services.is_empty() && file.services.keys().contains(&service))
+                    {
+                        Some((service, container))
+                    } else {
+                        None
+                    }
+                })
+        })
+        .into_group_map();
 
-        for service in services {
+    if args.quiet {
+        for container in containers.into_values().flatten() {
+            println!("{}", container.id);
+        }
+    } else if args.service {
+        for service in containers.into_keys() {
             println!("{service}");
         }
     } else {
-        let statuses = args
-            .status
-            .into_iter()
-            .map(|status| format!("status={status:?}").to_ascii_lowercase())
-            .collect::<Vec<_>>();
+        let filters = if containers.is_empty() {
+            vec![String::from("name=^$")]
+                .into_iter()
+                .chain(filters)
+                .collect::<Vec<_>>()
+        } else {
+            containers
+                .into_values()
+                .flatten()
+                .filter_map(|container| {
+                    container.names.front().map(|name| format!("name=^{name}$"))
+                })
+                .chain(filters)
+                .collect::<Vec<_>>()
+        };
 
         print!(
             "{}",
@@ -115,13 +124,15 @@ pub(crate) async fn run(args: Args, config: Config) -> Result<()> {
                     ["ps", "--filter", &format!("pod={name}")]
                         .into_iter()
                         .chain(if args.all { vec!["--all"] } else { vec![] })
-                        .chain(if args.format == Format::Json {
-                            vec!["--format", "json"]
-                        } else {
-                            vec![]
-                        })
-                        .chain(args.filter.iter().flat_map(|filter| ["--filter", filter]))
-                        .chain(statuses.iter().flat_map(|status| ["--filter", status]))
+                        .chain([
+                            "--format",
+                            match args.format {
+                                Format::Table =>
+                                    "table {{.Names}} {{.ID}} {{.Image}} {{.Command}} {{.CreatedHuman}} {{.Status}} {{.Ports}}",
+                                Format::Json => "json",
+                            }
+                        ])
+                        .chain(filters.iter().flat_map(|filter| ["--filter", filter]))
                 )
                 .await?
         );